backend/requirements.txt
tensorflow>=2.11.0
pandas>=1.5.0
numpy>=1.23.0
joblib>=1.2.0
psycopg2-binary>=2.9.0
python-dotenv>=0.21.0
scikit-learn>=1.0.0
gradio>=3.0.0
uvicorn>=0.20.0
fastapi>=0.95.0
pydantic>=1.10.0
groq>=0.4.0
schedule
